infra: the file dialog moves to the UI file rather than behind a header
widgetutils is header-only and includes <QFileDialog> itself, so calling a
wrapper from it left the translation unit needing Qt Widgets exactly as
before while taking the file off the audit - a number the audit's own header
says it must never report quietly. It also made all six files that include
widgetutils.h parse <QFileDialog>.

invokeFileDialog() goes to TLuaInterpreterUI.cpp instead, beside the other
Lua functions that drive the interface, so the file that is left really does
build without Qt Widgets and the count says so: it goes up by one, which is
the honest reading.

Found by Delwing in review.
