+++ src/TLuaInterpreterUI.cpp
@@ -68,6 +68,8 @@
 #include <QCollator>
 #include <QCoreApplication>
 #include <QDesktopServices>
+#include <QDir>
+#include <QFileDialog>
 #include <QFileInfo>
 #include <QMovie>
 #include <QVector>
@@ -4574,5 +4576,37 @@ int TLuaInterpreter::setCommandPulse(lua_State* L)
         return warnArgumentValue(L, __func__, error);
     }
     lua_pushboolean(L, success);
+    return 1;
+}
+
+// Documentation: https://wiki.mudlet.org/w/Manual:Lua_Functions#invokeFileDialog
+int TLuaInterpreter::invokeFileDialog(lua_State* L)
+{
+    const int n = lua_gettop(L);
+    if (!checkBoolArg(L, __func__, 1, "fileOrFolder") || !checkStringArg(L, __func__, 2, "dialogTitle") || (n > 2 && !checkStringArg(L, __func__, 3, "dialogLocation"))) {
+        return lua_error(L);
+    }
+
+    Host& host = getHostFromLua(L);
+    QString location = mudlet::getMudletPath(enums::profileHomePath, host.getName());
+    const bool luaDir = lua_toboolean(L, 1);
+    const QString title{lua_tostring(L, 2)};
+
+    if (n > 2) {
+        const QString target{lua_tostring(L, 3)};
+        const QDir dir(target);
+
+        if (dir.exists()) {
+            location = target;
+        }
+    }
+
+    if (!luaDir) {
+        const QString fileName = QFileDialog::getExistingDirectory(nullptr, title, location);
+        lua_pushstring(L, fileName.toUtf8().constData());
+        return 1;
+    }
+    const QString fileName = QFileDialog::getOpenFileName(nullptr, title, location);
+    lua_pushstring(L, fileName.toUtf8().constData());
     return 1;
 }
